api/requirements.txt
backend==0.2.4.1
fastapi==0.115.12
python-dotenv==1.1.0
uvicorn==0.34.2
